Add support for TPL & async / await
In order for Moq.Sequences to work correctly in the presence of TPL-
based asynchrony, the current approach of keeping track of the ambient
`Sequence` via `[ThreadStatic]` needs to be replaced, since `await`s
introduce an async boundary where the continuation might well be exe-
cuted on a different thread. For `async` / `await` to work correctly,
sequences need to "flow" with the call context.

This new behaviour directly violates Moq.Sequences' current semantics,
so it is designed as something that one must explicitly opt in to;
otherwise, the library defaults to the previous `[ThreadStatic]`-based
behavior.

Changing the library's behavior happens by setting the new static
property `Sequence.ContextMode = SequenceContextMode.Async`. (This
*might* become the default in the next major version of the library.)